indexer/pyproject.toml
[tool.poetry]
name = "indexer"
version = "0.1.0"
description = ""
authors = ["Maxime Duval <maximeduvalsy@gmail.com>"]
readme = "README.md"
include = [ "./src" ]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.95.2"
uvicorn = {extras = ["standard"], version = "^0.22.0"}
requests = "^2.31.0"
python-dotenv = "^1.0.0"
datasets = "^2.12.0"
pinecone-client = "^2.2.1"
sentence-transformers = "^2.2.2"
torch = "^2.0.1"
notebook = "^6.5.4"
faiss-cpu = "^1.7.4"
click = "^8.1.3"


[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
isort = "^5.12.0"
flake8 = "^6.0.0"
pyright = "^1.1.309"

[tool.isort]
profile = "black"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
